// src/transactions/utils/categories.enum.ts
export enum TransactionCategory {
  EXTERNAL_DEPOSIT = 'Dépôt Externe',
  SALES_REVENUE = 'Revenus des Ventes',
  LOAN = 'Emprunt Reçu',
  INVESTMENT = 'Investissement',
  GRANT = 'Subvention',
  BUYING = 'Achat',
  PAYING_FEES = 'Paiement de Frais',
  SALARY = 'Paiement des Salaires',
  RENT = 'Paiement du Loyer',
  UTILITIES = 'Factures (Électricité, Eau, Internet)',
  TRANSPORT = 'Transport',
  TAXES = 'Impôts',
  MAINTENANCE = 'Entretien et Réparations',
  DEBT_REPAYMENT = 'Remboursement de Dette',
  DONATION = 'Don'
}